Adjust setup for Kingfluencers
As tomorrow I start new job and will have to setup new machine, I've
decided that the simplest way to speed up this process is to simply
adjust scripts for this exact machine.

Fix for Python compilation was found on pyenv github issue[1] and
doesn't seem to be final solution. I couldn't really tell what to check
to ensure we're not installing again so I just check if it was an
upgrade or not to tell whether something changed.

[1] https://github.com/pyenv/pyenv/issues/1219